micro2/aula4/catalogo/requirements.test.txt
fastapi==0.100.0
pydantic==2.0.2
uvicorn==0.22.0
mysql-connector-python==8.0.33
SQLAlchemy==2.0.18
requests==2.31.0
boto3==1.26.155
pact-python==2.0.1
Unidecode==1.3.6
pytest-asyncio==0.21.1